Simplify CPU
In my initial run through, I missed that the jump and dest bits don't
really need much inspection beyond checking an individual bit. This
removed a bunch of unnecessary gyrations with DMuxes and such.

It's unclear to me how to order these HDL files. Given the complexity of
the CPU, I ordered it from MSB to LSB in the instruction and the parts
that tackle each. This allows the file to read from figuring out if it's
an A or C instruction, and conclude with setting up the program counter.